Bug fixes for object detection
This change includes important bug fixes for edge cases in the robot.detect_objects() method. The ._extrema() method was not including the first lidar scan point as a maximum if the distance starts flat or descending. The ._extrema_filt method was not including maxima that were near the start or end of the data set where the threshold could not be met because the end of the list was reached. Those maxima are important when scanning for objects with a small angle.